+++ tests/tests_client.py
@@ -77,7 +77,7 @@ class TestClient:
 
     @mock.patch("openweather_sdk.rest.geocoding._GeocodingAPI._direct")
     @mock.patch("openweather_sdk.rest.weather._WeatherAPI._get_current_wheather")
-    def test_get_location_weather(
+    def test_get_location_current_weather(
         self, mock_get_weather, mock_get_coordinates, weather_client
     ):
         with warnings.catch_warnings(record=True) as w:
@@ -98,7 +98,7 @@ class TestClient:
         mock_response_weather = mock.Mock()
         mock_response_weather.return_value = WEATHER_API_CORRECT_DATA
         mock_get_weather.side_effect = mock_response_weather
-        weather_data = weather_client._get_current_weather_coordinates(*coordinates)
+        weather_data = weather_client._get_current_weather(*coordinates)
         assert weather_data == WEATHER_API_CORRECT_DATA
 
         with pytest.raises(InvalidLocationException):
@@ -110,7 +110,7 @@ class TestClient:
 
     @mock.patch("openweather_sdk.rest.geocoding._GeocodingAPI._zip")
     @mock.patch("openweather_sdk.rest.weather._WeatherAPI._get_current_wheather")
-    def test_get_zip_weather(
+    def test_get_zip_current_weather(
         self, mock_get_weather, mock_get_coordinates, weather_client
     ):
         with warnings.catch_warnings(record=True) as w:
@@ -133,7 +133,7 @@ class TestClient:
         mock_response_weather = mock.Mock()
         mock_response_weather.return_value = WEATHER_API_CORRECT_DATA
         mock_get_weather.side_effect = mock_response_weather
-        weather_data = weather_client._get_current_weather_coordinates(*coordinates)
+        weather_data = weather_client._get_current_weather(*coordinates)
         assert weather_data == WEATHER_API_CORRECT_DATA
 
         with pytest.raises(InvalidLocationException):
@@ -173,7 +173,7 @@ class TestClient:
         mock_response_weather = mock.Mock()
         mock_response_weather.return_value = WEATHER_API_CORRECT_DATA
         mock_get_weather.side_effect = mock_response_weather
-        weather_data = weather_client._get_current_weather_coordinates(*coordinates)
+        weather_data = weather_client._get_current_weather(*coordinates)
         assert weather_data == WEATHER_API_CORRECT_DATA
 
         with pytest.raises(InvalidLocationException):
@@ -188,3 +188,37 @@ class TestClient:
         weather_client.remove()
         with pytest.raises(ClientDoesntExistException):
             weather_client.current_weather(zip_code="75000,FR")
+
+    @mock.patch("openweather_sdk.rest.geocoding._GeocodingAPI._direct")
+    @mock.patch("openweather_sdk.rest.forecast._ForecastAPI._get_5_day_forecast")
+    def test_forecast_5_day(
+        self, mock_get_weather, mock_get_coordinates, weather_client
+    ):
+        mock_response_coordinates = mock.Mock()
+        mock_response_coordinates.return_value = {
+            "name": "Paris",
+            "lat": 48.8588897,
+            "lon": 2.3200410217200766,
+        }
+        mock_get_coordinates.side_effect = mock_response_coordinates
+        coordinates = weather_client._get_location_coordinates("Paris")
+        assert coordinates == (2.32, 48.859)
+
+        mock_response_weather = mock.Mock()
+        mock_response_weather.return_value = WEATHER_API_CORRECT_DATA
+        mock_get_weather.side_effect = mock_response_weather
+        weather_data = weather_client._get_forecast_5_day(*coordinates)
+        assert weather_data == WEATHER_API_CORRECT_DATA
+
+        with pytest.raises(InvalidLocationException):
+            weather_client.forecast_5_day()
+
+        with pytest.raises(InvalidLocationException):
+            weather_client.forecast_5_day(zip_code=75000)
+
+        with pytest.raises(InvalidLocationException):
+            weather_client.forecast_5_day(location=42)
+
+        weather_client.remove()
+        with pytest.raises(ClientDoesntExistException):
+            weather_client.forecast_5_day(zip_code="75000,FR")