+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `forecast_5_day` method to enable 5 day forecast requests based on either
+  location name or zip/postal code.
+  
 ### Deprecated
 
 - `get_location_weather` method will be deprecated in version 1.0.0.